casting

+++ cicdwebapp/src/main/java/com/mytech/cicdwebapp/controller/HomeController.java
@@ -43,7 +43,9 @@ public class HomeController {
 			return fact;
 			
 		};		
-		return cal.getFactorial(no);
+		return (int)cal.getFactorial(no);
+		
+		
 	}
 	
 }
